Update deps, config and refactor TS config file
Now we have 3 TS config files: the default one is
in charge of generating type definitions, while the
others inherit from the base one and they are in
charge of generating different JS flavours.

Also, we now avoid to generage type definition twice:
before we were generating them both in the types and
in the esm folders. Now, all the type are in the types
folder only and the package.json exports for types
use that folder as the single source of truth.